requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
pywebview[edgehtml]
requests
mcp
openai
azure-identity